feat (shared): implement UuidToken (#107)
