+++ src/Shared/Domain/ValueObject/Token/UuidToken.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace App\Shared\Domain\ValueObject\Token;
+
+use App\Shared\Domain\ValueObject\Uuid\AbstractUuid;
+
+class UuidToken extends AbstractUuid
+{
+}
